+++ backend/career-guidance-backend/src/main/java/com/smartcareer/careerguidancebackend/repository/SkillAssessmentRepository.java
@@ -6,13 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
-public interface SkillAssessmentRepository extends JpaRepository<SkillAssessment, Integer> {
-
-    // ✅ Get all skill names for the student
+public interface SkillAssessmentRepository extends JpaRepository<SkillAssessment, Long> { // Change Integer to Long
     @Query("SELECT s.skillName FROM SkillAssessment s WHERE s.student = :student")
     List<String> findSkillsByStudent(StudentProfile student);
 
-    // ✅ Get all skill assessments for the student
     List<SkillAssessment> findByStudentOrderByAssessmentDateAsc(StudentProfile student);
-
 }
